Enable tests to be located everywhere
Being able to load tests from everywhere within the file tree makes it
possible to have test files located directly next to the files they are testing.
When creating new tests in the future, they can be created in the same
location with the code that needs to be tested. This makes it easier to
find related tests.